requirements.txt
streamlit
requests
fastapi
uvicorn
pydantic
python-multipart
langchain
langgraph
chromadb
langchain_huggingface
langchain
langchain-chroma
transformers
sentence-transformers
python-dotenv
asyncpg 
psycopg[binary]
sseclient
langchain-openai
langchain-ollama
pyang
langchain-community
langchain-core

